+++ greenfield/eldenring/data.py
@@ -5538,4 +5538,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:d8a2657706f02e932e6889438ccaa8006de51261ca46d1108371728a6f52aedc', 'module': 'data.py', 'body_sha256': 'sha256:5ba1a91442731be3e2169651be1524784f694b632e11fd8e1293053136dc90d2'}
+_GEN_STAMP = {'inputs_hash': 'sha256:e1fd94cf9217a10f81c7d4562611168d7128d73fdbc6a377126e15b9f37e857c', 'module': 'data.py', 'body_sha256': 'sha256:5ba1a91442731be3e2169651be1524784f694b632e11fd8e1293053136dc90d2'}